# wso2/apextick-api/openapi.json
{
  "components": {
    "schemas": {
      "AdminSeatResponse": {
        "properties": {
          "heldBy": {
            "type": "string"
          },
          "heldUntil": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "label": {
            "type": "string"
          },
          "sectionId": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "enum": [
              "AVAILABLE",
              "HELD",
              "BOOKED"
            ],
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "CurrentUser": {
        "properties": {
          "admin": {
            "type": "boolean"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "sub": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "EventDetailResponse": {
        "properties": {
          "availableSeats": {
            "format": "int64",
            "type": "integer"
          },
          "away": {
            "$ref": "#/components/schemas/TeamResponse"
          },
          "blurb": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "currencySymbol": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "fromPrice": {
            "type": "number"
          },
          "home": {
            "$ref": "#/components/schemas/TeamResponse"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "image": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "salesEndAt": {
            "format": "date-time",
            "type": "string"
          },
          "salesStartAt": {
            "format": "date-time",
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionResponse"
            },
            "type": "array"
          },
          "series": {
            "$ref": "#/components/schemas/SeriesResponse"
          },
          "seriesId": {
            "format": "int64",
            "type": "integer"
          },
          "seriesSlug": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "sport": {
            "type": "string"
          },
          "stadium": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tiers": {
            "items": {
              "$ref": "#/components/schemas/PriceTierResponse"
            },
            "type": "array"
          },
          "time": {
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "totalSeats": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "EventStatusRequest": {
        "properties": {
          "status": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "EventSummaryResponse": {
        "properties": {
          "availableSeats": {
            "format": "int64",
            "type": "integer"
          },
          "away": {
            "$ref": "#/components/schemas/TeamResponse"
          },
          "blurb": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "currencySymbol": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "fromPrice": {
            "type": "number"
          },
          "home": {
            "$ref": "#/components/schemas/TeamResponse"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "image": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "salesEndAt": {
            "format": "date-time",
            "type": "string"
          },
          "salesStartAt": {
            "format": "date-time",
            "type": "string"
          },
          "seriesId": {
            "format": "int64",
            "type": "integer"
          },
          "seriesSlug": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "sport": {
            "type": "string"
          },
          "stadium": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "time": {
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "totalSeats": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "EventUpsertRequest": {
        "properties": {
          "awayTeamId": {
            "format": "int64",
            "type": "integer"
          },
          "blurb": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "currency": {
            "minLength": 1,
            "type": "string"
          },
          "homeTeamId": {
            "format": "int64",
            "type": "integer"
          },
          "image": {
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "salesEndAt": {
            "format": "date-time",
            "type": "string"
          },
          "salesStartAt": {
            "format": "date-time",
            "type": "string"
          },
          "seriesId": {
            "format": "int64",
            "type": "integer"
          },
          "slug": {
            "minLength": 1,
            "type": "string"
          },
          "sport": {
            "minLength": 1,
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "venue": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "currency",
          "name",
          "slug",
          "sport",
          "startsAt",
          "venue"
        ],
        "type": "object"
      },
...
      "HoldRequest": {
        "properties": {
          "seatIds": {
            "items": {
              "format": "int64",
              "type": "integer"
            },
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "seatIds"
        ],
        "type": "object"
      },
      "HoldResponse": {
        "properties": {
          "eventId": {
            "format": "int64",
            "type": "integer"
          },
          "heldUntil": {
            "format": "date-time",
            "type": "string"
          },
          "holdSeconds": {
            "format": "int64",
            "type": "integer"
          },
          "seatIds": {
            "items": {
              "format": "int64",
              "type": "integer"
            },
            "type": "array"
          },
          "seats": {
            "items": {
              "$ref": "#/components/schemas/SeatResponse"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "LayoutRequest": {
        "properties": {
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionSpec"
            },
            "minItems": 1,
            "type": "array"
          },
          "tiers": {
            "items": {
              "$ref": "#/components/schemas/TierSpec"
            },
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "sections",
          "tiers"
        ],
        "type": "object"
      },
      "LayoutResult": {
        "properties": {
          "eventId": {
            "format": "int64",
            "type": "integer"
          },
          "seatsCreated": {
            "format": "int32",
            "type": "integer"
          },
          "sectionsCreated": {
            "format": "int32",
            "type": "integer"
          },
          "tiersCreated": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "PageResponseEventSummaryResponse": {
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/EventSummaryResponse"
            },
            "type": "array"
          },
          "last": {
            "type": "boolean"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "totalElements": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "PriceTierResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "perks": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "price": {
            "type": "number"
          },
          "remaining": {
            "format": "int64",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "SeatResponse": {
        "properties": {
          "col": {
            "format": "int32",
            "type": "integer"
          },
          "heldUntil": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "label": {
            "type": "string"
          },
          "mine": {
            "type": "boolean"
          },
          "price": {
            "type": "number"
          },
          "row": {
            "format": "int32",
            "type": "integer"
          },
          "sectionCode": {
            "type": "string"
          },
          "sectionId": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "enum": [
              "AVAILABLE",
              "HELD",
              "BOOKED"
            ],
            "type": "string"
          },
          "tierCode": {
            "type": "string"
          },
          "tierId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "SectionResponse": {
        "properties": {
          "available": {
            "format": "int64",
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "rows": {
            "format": "int32",
            "type": "integer"
          },
          "seatsPerRow": {
            "format": "int32",
            "type": "integer"
          },
          "side": {
            "type": "string"
          },
          "tierCode": {
            "type": "string"
          },
          "tierId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "SectionSpec": {
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "rows": {
            "format": "int32",
            "type": "integer"
          },
          "seatsPerRow": {
            "format": "int32",
            "type": "integer"
          },
          "side": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "tierCode": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "SeriesResponse": {
        "properties": {
          "blurb": {
            "type": "string"
          },
          "cities": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "currency": {
            "type": "string"
          },
          "currencySymbol": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "image": {
            "type": "string"
          },
          "kicker": {
            "type": "string"
          },
          "mobileHeroImage": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "scale": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "sport": {
            "type": "string"
          },
          "story": {
            "type": "string"
          },
          "tint": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "TeamResponse": {
        "properties": {
          "color": {
            "type": "string"
          },
          "flag": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "logo": {
            "type": "string"
          },
          "monogram": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "short": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "TierSpec": {
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "perks": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "price": {
            "type": "number"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for browsing events and holding/booking seats.",
    "title": "ApexTick Booking & Inventory API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/events": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "in": "query",
            "name": "sport",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "seriesId",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "series",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "latest",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseEventSummaryResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Browse every event, including drafts and cancellations",
        "tags": [
          "Admin: Events"
        ]
      },
      "post": {
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Admin: Events"
        ]
      }
    },
    "/admin/events/{id}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Admin: Events"
        ]
      },
      "get": {
        "operationId": "detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Event detail in any status, including drafts and cancellations",
        "tags": [
          "Admin: Events"
        ]
      },
      "put": {
        "operationId": "update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Admin: Events"
        ]
      }
    },
    "/admin/events/{id}/layout": {
      "post": {
        "operationId": "layout",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutResult"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Admin: Events"
        ]
      }
    },
...
    "/admin/events/{id}/status": {
      "patch": {
        "operationId": "setStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Admin: Events"
        ]
      }
    },
...
    "/admin/seats": {
      "get": {
        "operationId": "byEvent",
        "parameters": [
          {
            "in": "query",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminSeatResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Admin: Seats"
        ]
      }
    },
    "/admin/seats/{id}/release": {
      "post": {
        "operationId": "release_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "admin",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CurrentUser"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSeatResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Admin: Seats"
        ]
      }
    },
...
    "/events/{idOrSlug}": {
      "get": {
        "operationId": "detail_1",
        "parameters": [
          {
            "in": "path",
            "name": "idOrSlug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [],
        "summary": "Event detail with tiers and sections",
        "tags": [
          "Events"
        ],
        "x-auth-type": "None"
      }
    },
    "/events/{idOrSlug}/holds": {
      "delete": {
        "operationId": "release",
        "parameters": [
          {
            "in": "path",
            "name": "idOrSlug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CurrentUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Release all of the caller's holds for this event",
        "tags": [
          "Holds"
        ]
      },
      "post": {
        "operationId": "hold",
        "parameters": [
          {
            "in": "path",
            "name": "idOrSlug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CurrentUser"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HoldResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Hold up to N seats for the caller (all-or-nothing)",
        "tags": [
          "Holds"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}